add type level tests for reverse
these tests will check, on a type level, if the result of a function conforms to the expected type.
Note that the eventual "expect" call will always work - we just need _any_ usage of the result variable, otherwise, TypeScript will complain that this variable is unused.

This means you will now get a compile time error if the reverse function would return a different type for the given input that we expect here, and we will check this via the compile script that runs during testing

courtesy of https://2ality.com/2019/07/testing-static-types.html