docs: Disable single line breaks in docs site
So that JSDocs of functions can be formatted on multiple
lines without affecting the docs site. Double line breaks
will still be rendered as a new paragraph.

This is done by disabling the GFM (GitHub flavored markdown)
style of line breaks when parsing the function descriptions
with the 'marked' package.