docs: Move markdown processing to astro (#498)
Pull the markdown logic from our transform script and into astro. Now
that this code runs on the server we can do all the pre-processing
naturally as part of generating the site.

Using tailwind prose to support any kind of markdown that might come up,
this allows us to support tables and all sort of unexpected things.