feat(only): Add `only` (#510)
closes #467 

i don't think `only` can be made lazy; here's the argument. when called
with a value, we either:

* return it immediately as `next`. later if we get another value we
can't "take back" the value we gave before
* don't return it as `next`. but then we don't get called again, so
there's no way for the caller to get the value

there are changes that can be made to the architecture to allow it to be
lazy. the one that needs least changes, i think, is to change `pipe` to
be

```diff
    if ((lastLazySeq as any).single) {
-      ret = acc[0];
+      ret = acc[acc.length - 1];
    } else {
      ret = acc;
    }
```

this lets `only` "take back" a value it returned, and replace it with
`undefined` (and, surprisingly to me, all tests pass after this change).
i don't like it though

---

Make sure that you:

- [x] Typedoc added for new methods and updated for changed
- [x] Tests added for new methods and updated for changed
- [x] New methods added to `src/index.ts`
- [x] New methods added to `mapping.md`

---

<details><summary>We use semantic PR titles to automate the release
process!</summary>

https://conventionalcommits.org

PRs should be titled following using the format: `< TYPE >(< scope >)?:
description`

### Available Types:

- `feat`: new functions, and changes to a function's type that would
impact users.
- `fix`: changes to the runtime behavior of an existing function, or
refinements to it's type that shouldn't impact most users.
- `perf`: changes to function implementations that improve a functions
_runtime_ performance.
- `refactor`: changes to function implementations that are neither `fix`
nor `perf`
- `test`: tests-only changes (transparent to users of the function).
- `docs`: changes to the documentation of a function **or the
documentation site**.
- `build`, `ci`, `style`, `chore`, and `revert`: are only relevant for
the internals of the library.

For scope put the name of the function you are working on (either new or
existing).

</details>