feat: Code quality improvements to all utilities (via lint and manual inspection) (#555)
This PR complements previous work on adding additional lint rules with
the rules requiring manual fixes. This PR is long and risky. Most of the
changes are to types or slight code structure and not actual logic, but
some non-trivial changes appear in:
* _reduceLazy (bug surfaced: wasn't stopping iterations early when
iterator returned "done: true")
* equals (removed a lot of JSisms)
* pipe (refactored to not mutate function params directly).
* sample (bug surfaced via new types: would return the input array
instead of a copy)
* splitWhen (similar bug)

Other then these, the changes mainly made more stuff readonly, and
centralized all pipe/lazy-evaluation types into the pipe file so they
could be properly reused.

Notice the added `unknown` return value to the implementation function
(the one that calls purry) for all functions. There was no way to
configure the rule to ignore these functions but still check return
types for the actual implementation functions, which was what I wanted
to have lint check. With explicit return values we can now catch
implementation issues hidden by the fact that `purry` would "erase" the
types.