feat: isIncludedIn (#568)
A type-guard that checks existence in a set of values. Useful for
narrowing possible enum (literal union) values via a global const.

Optimizes dataFirst and dataLast implementations offering better
performance for each case.

Deprecated intersection and difference in favour of using filter in
combination with isIncludedIn so we can take advantage of the
performance gains and optimizations, and to make room for future
interestion and difference functions which will work more closely to the
mathematical definitions of multi-sets (which take into account
duplicate items, as could be present in arrays). See #368

---

Make sure that you:

- [x] Typedoc added for new methods and updated for changed
- [x] Tests added for new methods and updated for changed
- [x] New methods added to `src/index.ts`
- [ ] New methods added to `mapping.md`

---

<details><summary>We use semantic PR titles to automate the release
process!</summary>

https://conventionalcommits.org

PRs should be titled following using the format: `< TYPE >(< scope >)?:
description`

### Available Types:

- `feat`: new functions, and changes to a function's type that would
impact users.
- `fix`: changes to the runtime behavior of an existing function, or
refinements to it's type that shouldn't impact most users.
- `perf`: changes to function implementations that improve a functions
_runtime_ performance.
- `refactor`: changes to function implementations that are neither `fix`
nor `perf`
- `test`: tests-only changes (transparent to users of the function).
- `docs`: changes to the documentation of a function **or the
documentation site**.
- `build`, `ci`, `style`, `chore`, and `revert`: are only relevant for
the internals of the library.

For scope put the name of the function you are working on (either new or
existing).

</details>