feat(v2): deprecate and rename functions ahead of v2 release. (#571)
I am marking functions as deprecated in v1 ahead of removal in v2. This
includes functions that were sunset and functions that were just
renamed. For renamed toPairs and fromPairs this also includes adding the
dataLast impl back from pr #531 as the rename allows people to migrate
cleanly.

Notice that this PR is for v1, not v2!