fix(isIncludedIn): Only narrow on exact cases (#607)
Because of limitations of how typescript handles literal unions (and the
lack of ability to detect them in the typing system) certain cases
narrow the negated case too aggressively.

e.g. If our data is of type `1 | 2 | 3` and our container is `Array<1 |
2>` we can say that the data could be narrowed to `1 | 2`, but we can't
say that the negated case could be narrowed to `3` because the container
array could still be empty.

I couldn't find a way to tell Typescript to narrow only the successful
result without impacting the negative one, so we can only reduce the
"surface" for the narrowing variant of the function to those that make
the most sense. Even with this, there are erroneous cases, but at least
they are reduced to a very small subset.

Fixes: #603