fix(hasAtLeast): dont narrow on non-literal minimum (#618)
Initially I thought this fix would have to wait for v2 because I wanted
to disable the guard from working on non-literal numbers.

With the recent work I did for `isIncludedIn` I realized that we can
simply "turn off" the guard-ness for those cases.

With this solution this change could be considered a fix and not a
breaking change, so could be merged into v1.