feat(flat): flat replaces `flatten` and `flattenDeep` (#619)
ECMAScript 2019 added `Array.prototype.flat` which covers both cases. To
align with that function's naming and behaviour we deprecated the
existing functions and replaced them with a new implementation.

Because Remeda v1 doesn't support ES2019, we create our own
implementation of the function. In v2 we will use the built-in one
instead.

---

Make sure that you:

- [x] Typedoc added for new methods and updated for changed
- [x] Tests added for new methods and updated for changed
- [x] New methods added to `src/index.ts`
- [x] New methods added to `mapping.md`

---

<details><summary>We use semantic PR titles to automate the release
process!</summary>

https://conventionalcommits.org

PRs should be titled following using the format: `< TYPE >(< scope >)?:
description`

### Available Types:

- `feat`: new functions, and changes to a function's type that would
impact users.
- `fix`: changes to the runtime behavior of an existing function, or
refinements to it's type that shouldn't impact most users.
- `perf`: changes to function implementations that improve a functions
_runtime_ performance.
- `refactor`: changes to function implementations that are neither `fix`
nor `perf`
- `test`: tests-only changes (transparent to users of the function).
- `docs`: changes to the documentation of a function **or the
documentation site**.
- `build`, `ci`, `style`, `chore`, and `revert`: are only relevant for
the internals of the library.

For scope put the name of the function you are working on (either new or
existing).

</details>