fix(pickBy,omitBy): Don't partialize simple records (#706)
Closes: #696

When making a Record/indexed-object partial `undefined` is added to the
value type. This is incorrect when using exactOptionalProperties because
`undefined` is not a valid value and would never show up, making the
result type too wide.

By checking specifically for string and number keys (and their
combination) we can rebuild the output object differently for those
cases.