fix(join): Remove Joined type export (#718)
We don't provide types in our exports, only functions. It was only
exported because it was shared in 2 different functions.

The type itself was written before we had type-fest integrated into the
library, and is a copy of the Join type in type-fest (which I
contributed when I wrote the type), so I consolidates it's usage.