fix(hasSubObject): incorrect type narrowing in negate condition (#683)
Fixes: #678

The only way I come up to fix the issue with the negate condition in
case when the guard resolves to the same type as `data` is by using
something that will make the guard always different from `data`. One
solution can be by using a brand type. Brand type is an internal type
that doesn't affect anything but can help us identify the type result
better. The users of the library can safely ignore that brand type.

This is the first occurrence of a brand type so let me know if I need to
change its format/naming, etc.

Next, in a separate PR, I'm going to try to improve the overall typing
of `hasSubObject` to make the narrowing more smarter. I already did some
progress but stuck on making it work well with `pipe` function.

---

Make sure that you:

- [x] Typedoc added for new methods and updated for changed
- [x] Tests added for new methods and updated for changed
- [x] New methods added to `src/index.ts`
- [x] New methods added to `mapping.md`

---

<details><summary>We use semantic PR titles to automate the release
process!</summary>

https://conventionalcommits.org

PRs should be titled following using the format: `< TYPE >(< scope >)?:
description`

### Available Types:

- `feat`: new functions, and changes to a function's type that would
impact users.
- `fix`: changes to the runtime behavior of an existing function, or
refinements to it's type that shouldn't impact most users.
- `perf`: changes to function implementations that improve a functions
_runtime_ performance.
- `refactor`: changes to function implementations that are neither `fix`
nor `perf`
- `test`: tests-only changes (transparent to users of the function).
- `docs`: changes to the documentation of a function **or the
documentation site**.
- `build`, `ci`, `style`, `chore`, and `revert`: are only relevant for
the internals of the library.

For scope put the name of the function you are working on (either new or
existing).

</details>