fix(conditional): Use Object.assign to inject the defaultCase prop (#723)
Fixes: #720

Copying what I wrote in issue #720:

I managed to reproduce the issue. It bisects to #596, but that makes me
think that maybe that PR (that simply fixed a attw linit warning) just
surfaced the fact that our ESM build for this function was always
broken. I have a hunch that tsup (or esbuild) is not doing a good job
with `namespace` by using an IIFE to inject the property instead of
`Object.assign`, which is what I would expect would happen.