fix: key and value typing of unions of object types. (#728)
Closes #727.

I thought it was an inference issue so I tried to resolve it by
modifying the types and using `infer`. Moreover, I tried to keep using
the internal types `EnumerableStringKeyOf` and
`EnumerableStringKeyedValueOf`.

I would be happy to integrate any further adjustments as needed. 

---

Make sure that you:

- [ ] Typedoc added for new methods and updated for changed
- [x] Tests added for new methods and updated for changed
- [ ] New methods added to `src/index.ts`
- [ ] New methods added to `docs/src/pages/mapping.md`

---

<details><summary>We use semantic PR titles to automate the release
process!</summary>

https://conventionalcommits.org

PRs should be titled following using the format: `< TYPE >(< scope >)?:
description`

### Available Types:

- `feat`: new functions, and changes to a function's type that would
impact users.
- `fix`: changes to the runtime behavior of an existing function, or
refinements to it's type that shouldn't impact most users.
- `perf`: changes to function implementations that improve a functions
_runtime_ performance.
- `refactor`: changes to function implementations that are neither `fix`
nor `perf`
- `test`: tests-only changes (transparent to users of the function).
- `docs`: changes to the documentation of a function **or the
documentation site**.
- `build`, `ci`, `style`, `chore`, and `revert`: are only relevant for
the internals of the library.

For scope put the name of the function you are working on (either new or
existing).

</details>