feat(forEach): pass complex array shapes thru the pipe (#762)
The current implementation takes T as the array item type, and not the
array type itself; this prevents us from using the shape in the returned
type, which means we can't reconstruct the same shape on output.

This change is inline with how we treat array functions that return the
sampe shape (like `map`).

The problem with this typing scheme is that it breaks non-piped
("function factory") dataLast invocations, where previously the type
could be inferred from the param type of the callback (see test), but
this is OK because that isn't a use-case we care about in the lib (as it
doesn't seem very useful). This is the reason I called this a "feat" and
not a "fix" in the semantic commits convention, so that it bumps the
minor version and not just the patch version.