feat(times): Refined typing and improved perf (#772)
The current typing implementation doesn't use the provided size param to
build the output, this means that even for cases like times(5, ...) our
returned array type is not non-empty.

I also noticed that we are throwing on "invalid" inputs in a way which
doesn't conform with how we handle these kind of cases in other
functions. Instead, we should almost always fallback to a runtime
behaviour that makes sense, especially when a semanticly correct option
exists, and even more so when we can make the types work for it too.

Looking into this fix, I also wanted to try sparse arrays again, we have
a lint rule which might not make sense for us driving us away from
sparse arrays and it might not always be the right call. For this
function we can get performance benefits of about 2x-3x on large arrays
using sparse arrays.
I added a bunch of benchmarks to see what implementation is the best. I
will remove them before the merge because they won't make sense once
this PR is approved.