feat(shuffle): Maintain input shape (#781)
Currently we reconstruct the output from the input member types.

Instead we can use the input type itself to build the output. This
allows us to pass through qualities of the input like non-emptyness to
the output.

In other words, shuffle should have the same return type as sortBy
because it is effectively a sortBy a random number, so it can use the
same return type.