Tweaks annotated type error for lodash/fp
The error shown was exactly the same as Ramda, but in reality it is
slightly different. This doesn't seem like a big deal - I think the
point is still understood - but I thought while I was in here I might as
well fix it.

Here is a TS Playground for reference:

https://tinyurl.com/569s526w