feat(unique, uniqueWith, uniqueBy): Use input shape to improve return type. (#813)
The unique functions return a simple array, ignoring the shape of the
input. This means that some fidelity is lost when they are used in more
complex computations. We can use some basic characteristics of a deduped
array to improve the return type.

This isn't a perfect solution for all cases, as the test examples show,
because we don't dedupe stuff at the type level, we just build the
output based on the general character of the deduped array. This is
intentional, so that the same type could be used for all uniqueXXX
functions.