fix(takeXXX,dropXXX): Support unions of arrays/tuples (#789)
We still use the old style of typing array functions where T is the
member and not the array itself, this prevents the type from supporting
unions of arrays/tuples (`Array<number> | Array<string>`).

The typing for drop and take are still simple, and they don't take into
account the N itself, so they don't actually trim items from the
**type**, but that is a bigger change which might be breaking for some
flows. This change should be transparent to all existing solutions, and
it simply fixes an issue where our function doesn't even accept unions
in the input.