chore(deps): bump all in /docs (manual)  (#814)
Manually updating our packages in the docs project. This is done
manually as it's mostly updating eslint to v9.

The one plugin I couldn't manage to configure is `eslint-plugin-astro`.
I'm getting errors when it runs on astro files which makes me think it's
related to the recent release of the typescript-eslint plugin to v8. We
currently don't really have any lint running, so I prefer shipping with
out it, but with the rest of the eslint infra already set up with modern
versions.

We'll re-add the plugin in the near future when enough users encounter
issues with it and it gets fixed.