fix: updated lint rules -> nits to multiple utils (#851)
#801 should have been marked as a "fix" and not "chore" because it
touched and changed things that can impact users. For the same reason,
it should also have released a version.

This PR's primary goal is to cause a version bump and release, but it
also updates our rules with changes from eslint v9 and typescript-eslint
v8.

It also adds TypeScript 5.6 to our testing matrix.