chore(lint): add eslint-plugin-vitest (#865)
This plugin helps enforce higher-quality tests.

Having it enabled allows us to skip some of the comments we find
ourselves mentioning in code reviews often.

This PR changes a lot of files, mostly trivially via auto-fixes. The
interesting changes are in:
* doNothing.test
* indexBy.test-d
* Guard runtime checks: isArray, isBigInt, isBoolean, isDate, isDefined,
isError, isNonNull, isNonNullish, isNot, isNumber, isPromise, isString,
isSymbol, isTruthy
* pick.test had a really weird test from 4 years ago removed.

Note: this PR adds a lot of whitespace newlines. If possible, it's best
to review it with whitespaces ignored.

---------

Co-authored-by: Eran Hirsch <eran.ext@darrow.ai>