feat: add `toKebabCase` and `toSnakeCase` (#909)
By popular demand, adding simplified kebab case and snake case
functions.

The return types are trivial (`string`) because type-fest doesn't export
`SplitWords` and because its own KebabCase and SnakeCase types create
unions of possibilities instead of a concrete single outcome. #818 tries
to deal with that issue more concretely and can replace this
implementation once the changes in type-fest are approved.

I'm still deliberating with myself whether these functions need to be
part of the library or that we should export `splitWords` instead, as
they are simple compositions on top of it (with `join` and
`toLowerCase`). My issue with split words is that it isn't very smart;
it's a very mechanical way to define what a word is and not very
linguistic. By exporting it, we take responsibility for fixing all sorts
of issues people have with it when they use it for linguistic purposes.
On the other hand, the utilities that use it (snake_case, camelCase,
kebab-case) are mechanical themselves, and I don't think people will be
as surprised if they don't make sense linguistically.