fix(split): split on literal unions with separator as a prefix (#922)
A problem in type-fest's ArraySlice causes the issue
(https://github.com/sindresorhus/type-fest/issues/982) but could be
avoided in most cases by avoiding using ArraySlice when not needed.

This is a patch, and the problem would still exist when the optional
limit param is used, but for the vast majority of usages of this
utility, it should be enough to reduce the surface area of the issue,
giving time for the issue to be resolved on the type-fest side.

Closes: #921