fix(uniqueWith): return items that are not equal to themselves (#1001)
Fixes #999 

- Previous implementation checked each item for a match that existed at
a different index, skipping those. It checks the item's own equivalency
as part of the implementation, but this can unexpectedly remove items
due to interactions with the comparator's logic as shown in the issue.
- In the new implementation, the index equality check simultaneously
handles this edge case and reduces unnecessary calls of the comparator.

---

Make sure that you:

- [x] Typedoc added for new methods and updated for changed
- [x] Tests added for new methods and updated for changed
- [x] New methods added to `src/index.ts`
- [x] New methods added to `docs/src/content/mapping` for both Lodash
and Ramda.

---

<details><summary>We use semantic PR titles to automate the release
process!</summary>

https://conventionalcommits.org

PRs should be titled following using the format: `< TYPE >(< scope >)?:
description`

### Available Types:

- `feat`: new functions, and changes to a function's type that would
impact users.
- `fix`: changes to the runtime behavior of an existing function, or
refinements to it's type that shouldn't impact most users.
- `test`: tests-only changes (transparent to users of the function).
- `docs`: changes to the documentation of a function **or the
documentation site**.
- `build`, `ci`, `chore`, and `revert`: are only relevant for the
internals of the library.

For scope put the name of the function you are working on (either new or
existing).

</details>