docs: migrate functions data to Collections API (#997)
Redesigning the code that processes the function data coming from type
doc so that it's handled mostly by astro Content Collections API (v2).
This decouples the typedoc parsing from the rest of the system. ~~In the
future, it would allow us to create our own loader that would run
typedoc as an API instead of a utility and read the library's contents
dynamically, allowing HMR-like experience editing the docs directly via
the library source files~~ (Edit: **DONE**). Additionally, by removing
the transform file, we can move parsing closer to the components
themselves, allowing us to be more flexible with what and how we render
the docs.

---------

Co-authored-by: Eran Hirsch <eran.ext@darrow.ai>
Co-authored-by: Eran Hirsch <eranhirsch@DiggerM4C.local>