feat: monorepo (#1031)
Migrating the project to a monorepo structure and updating and
modernizing where it made sense.

I chose to use the most straightforward tools right now not to add
additional complexity when installing the project and contributing to
it. It wasn't clear to me that we needed this extra complexity. We can
consider using other package managers (pnpm) and other monorepo
automation tools (nx, lerna, turporepo) later on. On a personal note, I
found that people have a hard time installing and using pnpm when they
don't have prior experience with it and they usually end up doing things
the wrong way.

The CodeQL failure shown seen in the Checks section is real but is a
known issue and not related to this PR, it's only surfaced again as an
error because the file was moved. We can ignore it for this PR.

PS - I'm using `feat` instead of `chore` for this PR to trigger a minor
version bump. This is to allow us a clear point of reference for
bisecting in case errors are reported or issues are encountered after
release.