docs: tailwind v4 + react v19 (#1043)
Using the shadcn guide for tailwind 4:
https://ui.shadcn.com/docs/tailwind-v4. I recreated the whole project
config from scratch using the shadcn `init` and `add` CLI commands to
get a fresh start (instead of migrating the components individually).

This allows us to migrate to react 19 too, which is the last major
version upgrade from recent releases that we needed to sync up to.