fix(groupBy): Improve perf by using objects instead of Maps (#1080)
#1046 raised an issue with built-in props of the Object prototype
interfering with how we build our object.

#1049 fixed this using a Map (which doesn't have props and entries in
the same "namespace").

#1066 was an attempt to prefer objects over maps, but the perf didn't
justify it.

I still felt that using a Map was the wrong call here because it only
used the Map as an intermediate data structure and required recreating
the output once it was done. In this PR, I suggest a better
implementation that is significantly better performance-wise in almost
all cases. This was achieved after extensive testing of many other
similar implementations to pinpoint exactly what makes the engine purr
and what makes it stutter.

The final implementation is 2x to 8x faster on some important real-life
use cases and, in all cases, provides better perf than the existing
impl. There is only one case where the Map implementation is better, and
the difference is <x2, which is outshined by the perf benefits for the
other cases.

Once the PR is approved, I will remove the benchmark code, which will be
irrelevant going forward.