fix: add readme to package dist + other stuff (#1086)
Several things:
1. We don't have a readme shown in npmjs, this is since we moved to a
monorepo. For now i'm copying the readme inside the package as-is, but
we can either make a build step to copy it, or rewrite the readme to
better fit it's role in npmjs.
2. Changed TODOs to rely on package version and not a date, they won't
expire for regular development usage, but will expire for us when we
bump the major.
3. Removed the benchmark file, we don't need it anymore.