fix(hasAtLeast): handle more tuple/array shapes (#1085)
Reimplement the type using the new TupleParts. This includes fixes to
tuple parts and additional type tests for both.

The type itself was pulled out of the hasAtLeast utility because it
would be reused in the implementation of `groupByProp`.

I also refactored the RemedaErrorType to make it easier to use and to
use `never` as the underlaying type instead of a symbol so that types
that rely on it propagate the "neverness" of the error cases.