fix(filter): maintain tuple shape in result (#1087)
Using the new TupleParts we can improve the return type for filter when
used with a type-predicate. This allows us to consider every element in
the input tuple shape and consider whether it could possibly be in the
output or not.

Additionally we can also improve the result when filter is used with
trivial predicates (that always return true, or always return false),
allowing "short-circuit" style code where users effectively bypass the
filtering via the predicate itself.

The main type here was pulled out of the filter impl because it would be
the basis (together with ArrayRequiredPrefix) for the groupByProp return
type.