chore: publish to jsr (#1082)
It would be nice to be able to pull this package down from
[JSR](https://jsr.io/).
For me, JSR is much more ergonomic when it comes to ECMAScript Modules
and TypeScript.
This is especially nice for those developing with
[Deno](https://deno.com/) which has similar ergonomics.

Some additional configuration on their website would be needed to link
this repository there.
See
https://jsr.io/docs/publishing-packages#publishing-from-github-actions.

It seems there is already something registered at
https://jsr.io/@remeda/remeda.
So maybe this is already planned, but I didn't see anything related to
JSR in the repo.

See https://jsr.io/docs/why & https://jsr.io/docs/publishing-packages
for more info.

---------

Co-authored-by: Eran Hirsch <eranhirsch@gmail.com>