docs: hydration issues + a11y issues + perf optimizations (#1113)
* We have a version selector that allows users to switch the docs
between v1 and v2 at the top of the sidenav both on narrow ("mobile")
and wide ("desktop") breakpoints. The widget has some issues when
clicked on, a layout shift and a modality issue. These are probably
caused by the drop-down being implemented as a react portal that isn't
injected into the react tree properly by astro. It also adds additional
size to the bundle, although most users wouldn't need it or click it.

The selector is redundant because we also have an option in our top-nav
"Migration Guides" drop-down to switch to the v1 docs (and a separate
button to allow switching to the main docs).

We should drop the selector for now and reconsider the implementation in
the future if we get feedback or signal that it is needed.

* Fix a11y issues of our mobilve navbar sheet implementation surfaced by
radix in the console: adding a title and description.

* Fix hydration issues and optimize client load events to allow the
important parts to finish loading faster.