fix(sumBy,funnel,toCamelCase): fix test coverage issues (#1115)
Our coverage reporting was broken for a while, meaning we weren't
getting the signal when a PR broke coverage. Additionally, there have
been improvements to coverage in Vitest, which we can take advantage of.

This PR brings our coverage back to 100% statements, 100% branches, and
100% functions. It does so by slight changes to our impl and by adding
tests:

* funnel: we didn't have any tests without a reducer; this is an
important flow that needed proper testing. It also surfaced the fact
that our impl was lacking for this case. We now use a symbol to detect
the case of void function calls and make it more deliberate that the
functions are called without any arguments.
* toCamelCase: `words` would always return non-empty words, but we can't
represent that via TypeScript (because there is no non-empty string
type). Instead of shipping code that handles a case that can never
happen, we now use the erasable non-null assertion symbol.
* sumBy: relied on coalescing the "done" param, but there is no way to
cover this because it can't happen in v8 (probably in any engine); the
fact that "done" is optional is TypeScript playing it extremely safe, to
begin with. Instead of disabling coverage, I refactored it so that we
check the existence of done (which would always return true) and remove
the dud coalescing.

Test-only changes:
* removed our internal types from testing; they don't need to be
covered.
* added a test for withPrecision to validate that it works with
scientific notation inputs.