chore(tests): lint upgrade follow-ups + even greater standardization of our tests (#1135)
Following recent changes to the vitest eslint plugin it became apparent
that we should prefer importing vitest tools instead of using globals
(it also makes it easier for LLMs to provide higher quality suggestions
without requiring greater context).

Following that change, I also opted to standardize the usage of test/it
so that we don't start bikeshedding on this.

Finally cleaned up some other smaller rules.