feat(stringToPath): improve compatibility with Lodash (#1150)
Fixes: #779

stringToPath didn't handle numbers correctly for our use-case. Our
path-accepting utilities expect array access to be typed as `number` and
not as `${number}` which is what it previously returned.
Also we weren't handling certain edge-cases and complexities of parsing
a string path accuratly enough to how lodash does it.

I'm sending this out as a feat and not a fix so we bump the minor
version, as this would change runtime and typing behaviour for existing
call sites and can't be considered just a bug fix, even though it
probably only improves the types and surfaces existing bugs.